separated trips class into resources folder
